services/auth/requirements.txt
google-cloud-bigquery ~= 3.0.1
grpcio-tools ~= 1.44.0
grpc_interceptor ~= 0.13.1